refactor: move sync_(article|ticket) from temporal to lib
