+++ connectors/src/connectors/zendesk/lib/sync_article.ts
