+++ connectors/src/connectors/zendesk/lib/sync_ticket.ts
